+++ packages/graphback-codegen-resolvers/src/formatters/apollo/apolloResolverFormatter.ts
@@ -1,4 +1,4 @@
-import { ResolverGeneratorPluginOptions } from '../../plugins/ResolverGeneratorPlugin';
+import { ResolverGeneratorPluginOptions } from '../../plugin/ResolverGeneratorPlugin';
 import { createCustomResolversIndex as createCustomResolversIndexJS, createResolversIndexJS as createResolversIndexJS, resolverFileTemplateJS as resolverJSFileTemplate, rootResolversIndexJS } from './apolloJSResolverFormatter';
 import { createResolversIndexTS, resolverFileTemplateTS as resolverTSFileTemplate, rootResolversIndexTS } from './apolloTSResolverFormatter';
